Refactor sword classes into Swords namespace
Move melee sword implementations into a new RPGManagerLib.Items.Weapons.Melee.Swords namespace/folder and introduce an abstract Sword base there. Add BasicSword (internal), relocate BroadSword and GreatSword, and remove the old melee/Sword.cs. Update Weapon to include a SwordVariant Variant property and new constructor parameter (breaking change to Weapon signature), and add the new using in WeaponSaveData to match the moved types. Also update ROADMAP timestamp.
